requirements.txt
# Video Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Detection & Tracking
ultralytics>=8.0.0       
torch>=2.0.0
torchvision>=0.15.0

# Re-Identification
torchreid>=1.4.0          
faiss-cpu>=1.7.4          

# Graph Database
pymgclient>=1.3.0         
networkx>=3.1
neo4j>=5.0.0

# Backend API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
apscheduler>=3.10.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
loguru>=0.7.0

# Dashboard & Analytics
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
scipy>=1.11.0
pyvis>=0.3.2